Created modular tests for the franka robot so we can use them in other places. Also added ruff and changed some parts of test_utils.
