+++ dockerfiles/install_scripts/dev_packages.sh
@@ -12,6 +12,8 @@ pip install pytest-timeout
 # Specifying numpy range to work with ROS and Ultralytics:
 pip install "numpy>=1.23.0,<2.0"
 
+pip install ruff
+
 # Specifying currently newest version of transforms3d to avoid conflict with imported numpy.float in older version.
 pip install "transforms3d>=0.4.2"
