+++ ros2_ws/src/rcdt_franka/test/conftest.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: Alliander N. V.
+#
+# SPDX-License-Identifier: Apache-2.0
+
 from typing import Iterator
 
 import pytest
