+++ ros2_ws/src/rcdt_utilities/rcdt_utilities/test_utils.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: Alliander N. V.
+#
+# SPDX-License-Identifier: Apache-2.0
+
 import rclpy
 from builtin_interfaces.msg import Duration
 from control_msgs.action import FollowJointTrajectory
@@ -31,25 +35,26 @@ def get_joint_position(name: str) -> float:
     return position
 
 
-def call_trigger_service(node: rclpy.node.Node, service_name: str) -> bool:
+def call_trigger_service(node: rclpy.node.Node, service_name: str,timeout_server: float = 10.0, timeout_result: float = 30 ) -> bool:
     """Call a trigger service and return True if the service was called successfully."""
     client = node.create_client(Trigger, service_name)
-    if not client.wait_for_service(timeout_sec=5.0):
+    if not client.wait_for_service(timeout_sec=timeout_server):
         raise RuntimeError(f"Service {service_name} not available")
 
     future = client.call_async(Trigger.Request())
-    rclpy.spin_until_future_complete(node, future=future, timeout_sec=30)
+    rclpy.spin_until_future_complete(node, future=future, timeout_sec=timeout_result)
     return future.result() is not None
 
 
 def call_move_to_configuration_service(
-    node: rclpy.node.Node, configuration: str = "drop"
+    node: rclpy.node.Node, configuration: str = "drop", timeout_server: float = 10.0, timeout_result: float = 30
+
 ) -> bool:
     """Call the move_to_configuration service and return True if the service was called"""
     client = node.create_client(
         MoveToConfiguration, "/moveit_manager/move_to_configuration"
     )
-    if not client.wait_for_service(timeout_sec=5.0):
+    if not client.wait_for_service(timeout_sec=timeout_server):
         raise RuntimeError(
             "Service /moveit_manager/move_to_configuration not available"
         )
@@ -57,13 +62,14 @@ def call_move_to_configuration_service(
     request = MoveToConfiguration.Request()
     request.configuration = configuration
     future = client.call_async(request)
-    rclpy.spin_until_future_complete(node, future=future, timeout_sec=30)
+    rclpy.spin_until_future_complete(node, future=future, timeout_sec=timeout_result)
     return future.result() is not None
 
 
 def call_move_gripper_service(
     node: rclpy.node.Node, width: float, speed: float
 ) -> bool:
+    """Call the gripper to go to a defined width with a defined speed."""
     action_client = ActionClient(node, MoveGripper, "/fr3_gripper/move")
 
     if not action_client.wait_for_server(timeout_sec=5.0):
@@ -87,17 +93,20 @@ def call_move_gripper_service(
 
 
 def follow_joint_trajectory_goal(
-    singleton_node: rclpy.node.Node, positions: list[float]
+    singleton_node: rclpy.node.Node, positions: list[float],  controller: str = "fr3_arm_controller",
+    timeout_server: float = 10.0,
+    timeout_result: float = 30.0,
+    time_from_start: float = 3.0,
 ) -> bool:
     """Test sending a joint trajectory goal to the arm controller."""
 
     action_client = ActionClient(
         singleton_node,
         FollowJointTrajectory,
-        "/fr3_arm_controller/follow_joint_trajectory",
+        f"/{controller}/follow_joint_trajectory",
     )
 
-    if not action_client.wait_for_server(timeout_sec=5.0):
+    if not action_client.wait_for_server(timeout_sec=timeout_server):
         raise RuntimeError("Action server not available")
 
     goal_msg = FollowJointTrajectory.Goal()
@@ -113,16 +122,16 @@ def follow_joint_trajectory_goal(
 
     point = JointTrajectoryPoint()
     point.positions = positions
-    point.time_from_start = Duration(sec=3, nanosec=0)
+    point.time_from_start = Duration(sec=time_from_start, nanosec=0)
 
     goal_msg.trajectory.points.append(point)
 
     future = action_client.send_goal_async(goal_msg)
-    rclpy.spin_until_future_complete(singleton_node, future)
+    rclpy.spin_until_future_complete(singleton_node, future, timeout_sec=timeout_server)
     goal_handle = future.result()
     assert goal_handle.accepted
 
     result_future = goal_handle.get_result_async()
-    rclpy.spin_until_future_complete(singleton_node, result_future)
+    rclpy.spin_until_future_complete(singleton_node, result_future, timeout_sec=timeout_result)
     result = result_future.result().result
     return result
